bulk commit
