+++ src/display.cpp
@@ -9,7 +9,6 @@
 #include <sstream>
 
 #include "main.h"
-
 #include "opfunctions.h"
 
 lv_obj_t* scr = lv_obj_create(NULL, NULL);
