+++ src/main.cpp
@@ -6,116 +6,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-#include "main.h"
+#include <sstream>
 
+#include "main.h"
+#include "opcontrol.h"
 #include "opfunctions.h"
 
 // Controller Auton Indicator
-int scrcount = 1;
-bool ctrlScrBool = false;
-void ctrlrScr() {
-	std::string selAuton = arms::selector::b[abs(arms::selector::auton)];
-
-	if (!(scrcount % 25)) {
-		// Only print every 50ms, the controller text update rate is slow
-		if (ctrlScrBool == true) {
-			master.clear();
-			master.print(0, 0, "Auton: %s", selAuton.c_str());
-			printf("auton log");
-			ctrlScrBool = !ctrlScrBool;
-		} else {
-			master.clear();
-			master.print(0, 0, "Brake: %s", (pbrake ? "ON" : "OFF"));
-			printf("brake log");
-			ctrlScrBool = !ctrlScrBool;
-		}
-	}
-
-	scrcount++;
-	pros::delay(200);
-}
-
-pros::task_t matchTimerTask = (pros::task_t)NULL;
-
-// Match Timer Indicator
-int matchTimerCount = 105;
-void matchTimer() {
-	printf("Match Timer: %d\n", matchTimerCount);
-	pros::delay(1000);
-
-	while (true) {
-		if (matchTimerCount == 1) { // End of match
-			master.rumble("----");
-			pros::delay(1);
-			if (matchTimerTask) {
-				pros::Task(matchTimerTask).remove();
-				matchTimerTask = (pros::task_t)NULL;
-
-				matchTimerCount = 105;
-
-				printf("Match Timer Task Removed\n");
-			}
-		} else if (matchTimerCount == 35) { // 75 seconds into Driver Control
-			master.rumble(".-.-.");
-		} else if (matchTimerCount == 60) { // 60 seconds into Driver Control
-			master.rumble(". .");
-		} else if (matchTimerCount == 75) { // 45 seconds into Driver Control
-			master.rumble("-");
-		}
-
-		matchTimerCount--;
-		printf("Match Timer: %d\n", matchTimerCount);
-		pros::delay(1000);
-	}
-}
-
-/*
- * Runs the operator control code. This function will be started in its own task
- * with the default priority and stack size whenever the robot is enabled via
- * the Field Management System or the VEX Competition Switch in the operator
- * control mode.
- *
- * If no competition control is connected, this function will run immediately
- * following initialize().
- *
- * If the robot is disabled or communications is lost, the
- * operator control task will be stopped. Re-enabling the robot will restart the
- * task, not resume it from where it left off.
- */
-void opcontrol() {
-	if (pros::competition::is_connected()) {
-		matchTimerTask = pros::Task(matchTimer);
-	}
-
-	// Run Loop
-	while (true) {
-		// Steering
-		arms::chassis::arcade(
-			master.get_analog(ANALOG_LEFT_Y) * (double)100 / 127,
-		    master.get_analog(ANALOG_RIGHT_X) * (double)100 /127
-		);
-
-		// Arm
-		rightLift.move_velocity(master.get_analog(ANALOG_RIGHT_Y) * (double)100 / 127);
-		leftLift.move_velocity(master.get_analog(ANALOG_RIGHT_Y) * (double)100 / 127);
-		// Game Controls
-		gameSystemControls();
-
-		// Brake System
-		// Uses basic logic for toggle button
-		if (master.get_digital_new_press(DIGITAL_A) == 1) {
-			pbrake = !pbrake;
-		}
-		prosBrake(pbrake);
-
-		if (master.get_digital_new_press(DIGITAL_X) &&
-		    !pros::competition::is_connected())
-			autonomous();
-
-		// Lastly, delay
-		pros::delay(10);
-	}
-}
+std::string selAuton;
 
 /**
  * Runs initialization code. This occurs as soon as the program is started.
@@ -126,40 +24,53 @@ void opcontrol() {
 void initialize() {
 	// ARMS & Controller init and reset IMU sensor
 	arms::chassis::init();
-	arms::odom::init();
 	arms::pid::init();
-	arms::chassis::imu.reset();
 
-	pros::Task controllerTask(ctrlrScr);
-	pros::Task task{[=] {
+	pros::Task controllerTask{[=] {
+		master.clear();
+
 		while (true) {
-			printf("Task Count: %3d\n", pros::Task::get_count());
-            pros::delay(1000);
+			// Only print every 50ms, the controller text update rate is slow
+			selAuton = arms::selector::b[abs(arms::selector::auton)];
+
+			std::stringstream autonstr;
+			autonstr << "Auton: " << arms::selector::auton << "\r";
+			std::stringstream brakestr;
+			brakestr << "Brake: " << (pbrake ? "ON" : "OFF") << "\r";
+
+			master.print(0, 0, autonstr.str().c_str());
+			pros::delay(50);
+			master.print(1, 0, brakestr.str().c_str());
+			pros::delay(50);
+			if (pros::competition::is_connected()) {
+				master.print(2, 0, "Match Timer: %d\r", matchTimerCount);
+			} else {
+				// master.print(2, 0, "Task Count: %d\r", pros::Task::get_count());
+				master.print(2, 0, "Gyro: %f\r", imu_sensor.get_heading());
+			}
+
+			pros::delay(50);
 		}
-    }};
+	}};
+
+	// Task counter Task
+	// pros::Task taskTask{[=] {
+	// 	// while (true) {
+	// 	// 	printf("Task Count: %d\n", pros::Task::get_count());
+	// 	// 	pros::delay(2000);
+	// 	// }
+	// }};
 
 	// Set display
-	if (!pros::competition::is_connected()) display();
+	// if (!pros::competition::is_connected()) display();
+	arms::selector::init();
 
 	// Set brakes on to active bold
-	rightLift.set_brake_mode(pros::E_MOTOR_BRAKE_BRAKE);
-	leftLift.set_brake_mode(pros::E_MOTOR_BRAKE_BRAKE);
+	// E_MOTOR_BRAKE_BRAKE
+	rightLift.set_brake_mode(pros::E_MOTOR_BRAKE_HOLD);
+	leftLift.set_brake_mode(pros::E_MOTOR_BRAKE_HOLD);
 	clawM.set_brake_mode(pros::E_MOTOR_BRAKE_HOLD);
 	winchM.set_brake_mode(pros::E_MOTOR_BRAKE_HOLD);
-}
-
-void killTask() {
-	if (matchTimerTask) {
-		pros::Task(matchTimerTask).remove();
-		matchTimerTask = (pros::task_t)NULL;
-
-		matchTimerCount = 105;
-
-		printf("Match Timer Task Removed\n");
-	}
-
-	// small delay to allow tasks to be removed from run queue
-	pros::delay(10);
 }
 
 /**
@@ -172,7 +83,7 @@ void disabled() {
 	printf("Disabled called %d\n", count++);
 
 	// kill any tasks we may have started and do not need now
-	killTask();
+	// killTask();
 
 	// disabled is actually a task as well
 	// we can either return or block here doing something useful
@@ -193,10 +104,8 @@ void disabled() {
  */
 void competition_initialize() {
 	static int count = 1;
-    printf("Comp Init called %d", count++);
+	printf("Comp Init called %d\n", count++);
 
-    // if cable is removed and then attached we may need to
-    killTask();
-
-	arms::selector::init();
+	// if cable is removed and then attached we may need to
+	killTask();
 }
